+++ smartcontract/programs/doublezero-serviceability/src/processors/device/interface/activate.rs
@@ -69,7 +69,7 @@ pub fn process_activate_device_interface(
         .find_interface(&value.name)
         .map_err(|_| DoubleZeroError::InterfaceNotFound)?;
 
-    if iface.status == InterfaceStatus::Deleting {
+    if iface.status != InterfaceStatus::Pending && iface.status != InterfaceStatus::Unlinked {
         return Err(DoubleZeroError::InvalidStatus.into());
     }